% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\docType{data}
\name{citation_main}
\alias{citation_main}
\title{citation_main - citation main panel}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
citation_main
}
\description{
citation_main - citation main panel
}
\keyword{datasets}
