% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\docType{data}
\name{citation_side}
\alias{citation_side}
\title{citation_side - citation_side panel}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
citation_side
}
\description{
citation_side - citation_side panel
}
\keyword{datasets}
