% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{fileFormats}
\alias{fileFormats}
\title{fileFormats - available formats}
\format{
An object of class \code{character} of length 8.
}
\usage{
fileFormats
}
\description{
fileFormats - available formats
}
\keyword{datasets}
