% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{input.options.default}
\alias{input.options.default}
\title{input.options.default - default input file option}
\format{
An object of class \code{character} of length 1.
}
\usage{
input.options.default
}
\description{
input.options.default - default input file option
}
\keyword{datasets}
