% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{input.options}
\alias{input.options}
\title{input.options - input file options}
\format{
An object of class \code{list} of length 3.
}
\usage{
input.options
}
\description{
input.options - input file options
}
\keyword{datasets}
