% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\docType{data}
\name{load_data_content_side}
\alias{load_data_content_side}
\title{load_data_content_side  - load data side panel}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
load_data_content_side
}
\description{
load_data_content_side  - load data side panel
}
\keyword{datasets}
