% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{method.options.all}
\alias{method.options.all}
\title{method.options.all - all method options}
\format{
An object of class \code{list} of length 8.
}
\usage{
method.options.all
}
\description{
method.options.all - all method options
}
\keyword{datasets}
