% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{method.options.default}
\alias{method.options.default}
\title{method.options.default - default method options}
\format{
An object of class \code{character} of length 1.
}
\usage{
method.options.default
}
\description{
method.options.default - default method options
}
\keyword{datasets}
