% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{method.options}
\alias{method.options}
\title{method.options - available method options}
\format{
An object of class \code{list} of length 4.
}
\usage{
method.options
}
\description{
method.options - available method options
}
\keyword{datasets}
