% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{significance.options.default}
\alias{significance.options.default}
\title{significance.options.default - default significance option}
\format{
An object of class \code{list} of length 1.
}
\usage{
significance.options.default
}
\description{
significance.options.default - default significance option
}
\keyword{datasets}
