% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{significance.options}
\alias{significance.options}
\title{significance.options - available significance options}
\format{
An object of class \code{list} of length 10.
}
\usage{
significance.options
}
\description{
significance.options - available significance options
}
\keyword{datasets}
