% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\docType{data}
\name{top_part}
\alias{top_part}
\title{top_part - top navigation bar}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
top_part
}
\description{
top_part - top navigation bar
}
\keyword{datasets}
